Return Decimal value for numeric search for issue1805

+++ tryton/gui/window/view_form/widget_search/form.py
@@ -298,7 +298,7 @@ WIDGETS_TYPE = {
     'date': (calendar.Calendar, 2),
     'datetime': (calendar.Calendar, 2),
     'float': (float.Float, 2),
-    'numeric': (float.Float, 2),
+    'numeric': (float.Numeric, 2),
     'integer': (integer.Integer, 2),
     'biginteger': (integer.Integer, 2),
     'selection': (selection.Selection, 2),
